More updates to docs
