+++ docs/core_docs/docs/integrations/chat/google_palm.mdx
@@ -1,5 +1,6 @@
 ---
 sidebar_label: Google PaLM
+sidebar_class_name: hidden
 ---
 
 import CodeBlock from "@theme/CodeBlock";
