+++ docs/core_docs/docs/integrations/text_embedding/google_generativeai.mdx
@@ -2,13 +2,18 @@ import CodeBlock from "@theme/CodeBlock";
 
 # Google Generative AI
 
-// Docs here
+You can access Google's generative AI embeddings models through
+`@langchain/google-genai` integration package.
+
+Get an API key here: https://ai.google.dev/tutorials/setup
+
+You'll need to install the `@langchain/google-genai` package:
 
 ```bash npm2yarn
 npm install @langchain/google-genai
 ```
 
-// Docs here
+## Usage
 
 import GoogleGenerativeAIExample from "@examples/models/embeddings/googlegenerativeai.ts";
