+++ docs/core_docs/docs/integrations/text_embedding/google_palm.mdx
@@ -1,3 +1,7 @@
+---
+sidebar_class_name: hidden
+---
+
 import CodeBlock from "@theme/CodeBlock";
 
 # Google PaLM
